some warning fixes, changed the_addition and the_subtraction to floats instead of int

+++ SISASG/Source/Bullet.cpp
@@ -35,8 +35,8 @@ void Bullet::Update(double dt)
     {
         spaceship spaceship1;
 
-        this->position += view * dt * this->bbSpeed;
-        this->timeAlive += dt;
+        this->position += view * (float)dt * (float)this->bbSpeed;
+        this->timeAlive += (float)dt;
 
         if (this->timeAlive > 0.5)
         {
